pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quran-lexicon-pipeline"
version = "0.1.0"
description = "SQLite Quran lexicon pipeline, Tkinter explorer, and question-refiner LLM workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "networkx>=3.2.0",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",
    "pyarabic>=0.6.15",
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.7",
    "python-docx>=1.1.0",
    "xhtml2pdf>=0.2.15",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
